RemotePlay-based Connection (#18)
* Scaffold out RemotePlayDeviceConnection

* Scaffold out "session" handling

* Add initial session init nonce fetching

* Prepare CryptoStrategy for use with sessions

* Add Codec generation methods for auth uses (IE: for sessions)

* Add some initial (untested) session connection logic

I think we ought to be able to reuse the TcpDeviceSocket logic for
sending and receiving packets...

* Refactor RemotePlaySession to use TcpDeviceSocket

* Fix missed import update

* Try to send some extra params

* Fix endianness of counter; fix LegacyCryptoStrategy algorithm

* Tweak how registKey gets encoded

* Pad registkey to 16 bytes

* Verify RP-Key length

* Unify RP-Key padding

Probably unnecessary since it *should* be 16 bytes

* Fix: not completely decrypting RemotePlay responses

In particular, this resulted in us getting an incomplete RP-Key, which
was the cause of our consistently failed auth attempts at the ctrl
endpoint

* Refactor Modern and Legacy CryptoStrategies to share a base class

This simplifies some code sharing, and gives us one spot to validate
RP-Key

* Experiment with LOGIN command

Not quite sure what's going on...

* Fix socket management for RemotePlay

We have to ref() it to prevent the agent from killing things

* Make `options` optional in TcpDeviceSocket constructor

Fixes tests

* Refactor login flow into a Proc

* Check login response by parsing it into a proper packet

* Begin to prepare support for passcode auth

* Fix missing protocol bits for passcode auth

* Respond to Heartbeat messages automatically

* Provide the passcode, if we have it, on request

* Remove unnecessary RemotePlaySession abstraction

Also cleans up RemotePlayDeviceConnection somewhat

* Clean up very loud logging

* Fix endianness of outgoing packet type

* Further simplify RemotePlayDeviceConnection

Standby now works properly for PS5!

* Add logging to LoginProc

* Cleanup error suppression in login command

* Fix byte ordering of packet length, etc.

* Fix: not using connectionConfig in openConnection from login command

* Fix excess padding

* Don't "swap" codec for RemotePlay session; use it as initial

* Ensure we *only* decode the *current* packet for RemotePlayPacketCodec

The BufferPacketProcessor sort of assumes that the whole stream is
encoded, but that's not how the RemotePlay protocol works.

* Let the protocol handle incoming packets before receivers

Just in case order is significant

* Don't try to encode/decode an empty buffer

In the RemotePlay protocol, we should *only* increment the counters if
there was actually a payload to encrypt/decrypt. We were *always* trying
to encrypt/decrypt, which incorrectly incremented counters, causing a
desync with the host device.

Passcode auth works now!